+++ src/gameserver/core/framework/LayerManager/AgentLayerManager/IAgentManager.ts
@@ -1,0 +1,9 @@
+import { IPostWorldUpdateListener } from './IPostWorldUpdateListener';
+import { IPreWorldUpdateListener } from './IPreWorldUpdateListener';
+import { IFrameMetricProvider } from '../../GameLoopRunner';
+
+export interface IAgentManager
+    extends
+    Partial<IPostWorldUpdateListener>,
+    Partial<IPreWorldUpdateListener> {
+}