+++ src/gameserver/core/framework/LayerManager/AgentLayerManager/IPostWorldUpdateListener.ts
@@ -1,0 +1,5 @@
+import { IFrameMetricProvider } from '../../GameLoopRunner';
+
+export interface IPostWorldUpdateListener {
+    onPostWorldUpdate(frameMetricProvider: IFrameMetricProvider): void;
+}