+++ src/gameserver/core/framework/LayerManager/AgentLayerManager/IPreWorldUpdateListener.ts
@@ -1,0 +1,5 @@
+import { IFrameMetricProvider } from '../../GameLoopRunner';
+
+export interface IPreWorldUpdateListener {
+    onPreWorldUpdate(frameMetricProvider: IFrameMetricProvider): void;
+}