+++ src/gameserver/core/framework/LayerManager/AgentLayerManager/index.ts
@@ -1,0 +1,2 @@
+export { AgentLayerManager } from './AgentLayerManager';
+export { IAgentManager } from './IAgentManager';