+++ src/gameserver/core/framework/LayerManager/ILayerManager.ts
@@ -1,10 +1,10 @@
 import { TypedIdentifier } from '../TypedIdentifier';
 
-export interface ILayerManager {
+export interface ILayerManager<T> {
     initModules(): void;
 
-    registerModule<T>(
-        create: (locate: <J>(identifier: TypedIdentifier<J>) => J) => T,
-        ...identifiers: TypedIdentifier<T>[]
+    registerModule<J extends T>(
+        create: (locate: <K>(identifier: TypedIdentifier<K>) => K) => J,
+        ...identifiers: TypedIdentifier<J>[]
     ): void;
 }