+++ src/gameserver/core/framework/LayerManager/index.ts
@@ -1,3 +1,4 @@
 export { ILayerManager } from './ILayerManager';
 export { LayerManager } from './LayerManager';
 export { IUpdatableSystem, SystemManager } from './SystemManager';
+export { IAgentManager, AgentLayerManager } from './AgentLayerManager';