+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - GameServer: `GameLoopRunner` class
 - Change tsconfig module type to `commonjs` and output all files, not bundle
 - Add method binds to `DependencyLocator`
+- Add `AgentLayerManager` class
 
 ### Changed
 ### Deprecated